Create the compose function to fetch the viewer info
