+++ client/package.json
@@ -31,7 +31,8 @@
     "react-navigation-collapsible": "^3.0.3",
     "react-navigation-drawer": "^2.3.3",
     "react-navigation-stack": "^1.10.3",
-    "react-navigation-tabs": "^2.6.2"
+    "react-navigation-tabs": "^2.6.2",
+    "recompose": "^0.30.0"
   },
   "devDependencies": {
     "@babel/core": "^7.7.4",
